Fix: Рефакторит маршрутизатор
