+++ src/components/app/app.tsx
@@ -1,15 +1,11 @@
 import {BrowserRouter} from 'react-router-dom';
-import RoutesGenerator from '../../routes-generator/routes-generator.tsx';
-import {PageProps} from '../../utils/types.ts';
-
-type AppProps = {
-  pageProps: PageProps;
-};
+import AppRouter from '../../app-router/app-router.tsx';
+import {AppProps} from '../../utils/types.ts';
 
 export default function App({pageProps}: AppProps) {
   return(
     <BrowserRouter>
-      {RoutesGenerator.generateRoutes(pageProps)}
+      <AppRouter pageProps={pageProps}/>
     </BrowserRouter>
   );
 }
