+++ src/utils/types.ts
@@ -5,13 +5,17 @@ export type PageProps = {
   reviews: Review[];
   favoritesObject: FavoritesObject;
   favoritesQuantity: number;
-}
+};
+
+export type AppProps = {
+  pageProps: PageProps;
+};
 
 type Location = {
   latitude: number;
   longitude: number;
   zoom: number;
-}
+};
 
 type OfferScaffolding = {
   id: string;
